+++ daskms/experimental/zarr/__init__.py
@@ -1,3 +1,5 @@
+from functools import reduce
+from operator import mul
 from pathlib import Path
 
 import dask
@@ -81,6 +83,23 @@ def create_array(ds_group, column, column_schema,
         chunks = column_schema.chunks
 
     zchunks = zarr_chunks(column, column_schema.dims, chunks)
+
+    if column_schema.dtype == object:
+        if reduce(mul, zchunks, 32) >= 2**(32 - 1):
+            raise ValueError(
+                f"Column {column} has an object dtype. "
+                f"Given an estimate of 32 bytes per entry "
+                f"the chunk of dimensions {zchunks}"
+                f"may exceed zarr's 2GiB chunk limit"
+            )
+    else:
+        size = np.dtype(column_schema.dtype).itemsize
+        if reduce(mul, zchunks, size) >= 2**(32 - 1):
+            raise ValueError(
+                f"Column {column} has a chunk of "
+                f"dimension {zchunks} that will exceed "
+                f"zarr's 2GiB chunk limit"
+            )
 
     array = ds_group.require_dataset(column, column_schema.shape,
                                      chunks=zchunks,